--- a/distsql/request_builder.go
+++ b/distsql/request_builder.go
@@ -14,6 +14,7 @@
 package distsql
 
 import (
+	"fmt"
 	"math"
 
 	"github.com/pingcap/parser/mysql"
@@ -43,7 +44,7 @@ func (builder *RequestBuilder) Build() (*kv.Request, error) {
 }
 
 // SetMemTracker sets a memTracker for this request.
-func (builder *RequestBuilder) SetMemTracker(sctx sessionctx.Context, label string) *RequestBuilder {
+func (builder *RequestBuilder) SetMemTracker(sctx sessionctx.Context, label fmt.Stringer) *RequestBuilder {
 	t := memory.NewTracker(label, sctx.GetSessionVars().MemQuotaDistSQL)
 	t.AttachTo(sctx.GetSessionVars().StmtCtx.MemTracker)
 	builder.Request.MemTracker = t