--- a/executor/join.go
+++ b/executor/join.go
@@ -15,6 +15,7 @@ package executor
 
 import (
 	"context"
+	"fmt"
 	"sync"
 	"sync/atomic"
 	"time"
@@ -262,12 +263,14 @@ func (e *HashJoinExec) wait4Inner() (finished bool, err error) {
 	return false, nil
 }
 
+var innerResultLabel fmt.Stringer = stringutil.StringerStr("innerResult")
+
 // fetchInnerRows fetches all rows from inner executor, and append them to
 // e.innerResult.
 func (e *HashJoinExec) fetchInnerRows(ctx context.Context) error {
 	e.innerResult = chunk.NewList(e.innerExec.retTypes(), e.initCap, e.maxChunkSize)
 	e.innerResult.GetMemTracker().AttachTo(e.memTracker)
-	e.innerResult.GetMemTracker().SetLabel(stringutil.StringerStr("innerResult"))
+	e.innerResult.GetMemTracker().SetLabel(innerResultLabel)
 	var err error
 	for {
 		if e.finished.Load().(bool) {
@@ -612,6 +615,8 @@ func (e *NestedLoopApplyExec) Close() error {
 	return e.outerExec.Close()
 }
 
+var innerListLabel fmt.Stringer = stringutil.StringerStr("innerList")
+
 // Open implements the Executor interface.
 func (e *NestedLoopApplyExec) Open(ctx context.Context) error {
 	err := e.outerExec.Open(ctx)
@@ -627,7 +632,7 @@ func (e *NestedLoopApplyExec) Open(ctx context.Context) error {
 	e.memTracker = memory.NewTracker(e.id, e.ctx.GetSessionVars().MemQuotaNestedLoopApply)
 	e.memTracker.AttachTo(e.ctx.GetSessionVars().StmtCtx.MemTracker)
 
-	e.innerList.GetMemTracker().SetLabel(stringutil.StringerStr("innerList"))
+	e.innerList.GetMemTracker().SetLabel(innerListLabel)
 	e.innerList.GetMemTracker().AttachTo(e.memTracker)
 
 	return nil