--- a/executor/load_data.go
+++ b/executor/load_data.go
@@ -26,6 +26,7 @@ import (
 	"github.com/pingcap/tidb/types"
 	"github.com/pingcap/tidb/util/chunk"
 	"github.com/pingcap/tidb/util/logutil"
+	typ "github.com/pingcap/tidb/util/types"
 	"go.uber.org/zap"
 )
 
@@ -39,7 +40,7 @@ type LoadDataExec struct {
 
 // NewLoadDataInfo returns a LoadDataInfo structure, and it's only used for tests now.
 func NewLoadDataInfo(ctx sessionctx.Context, row []types.Datum, tbl table.Table, cols []*table.Column) *LoadDataInfo {
-	insertVal := &InsertValues{baseExecutor: newBaseExecutor(ctx, nil, "InsertValues"), Table: tbl}
+	insertVal := &InsertValues{baseExecutor: newBaseExecutor(ctx, nil, typ.InstantStr("InsertValues")), Table: tbl}
 	return &LoadDataInfo{
 		row:          row,
 		InsertValues: insertVal,