--- a/executor/prepared.go
+++ b/executor/prepared.go
@@ -15,6 +15,7 @@ package executor
 
 import (
 	"context"
+	"github.com/pingcap/tidb/util/types"
 	"math"
 	"sort"
 
@@ -87,7 +88,7 @@ type PrepareExec struct {
 
 // NewPrepareExec creates a new PrepareExec.
 func NewPrepareExec(ctx sessionctx.Context, is infoschema.InfoSchema, sqlTxt string) *PrepareExec {
-	base := newBaseExecutor(ctx, nil, "PrepareStmt")
+	base := newBaseExecutor(ctx, nil, types.InstantStr("PrepareStmt"))
 	base.initCap = chunk.ZeroCapacity
 	return &PrepareExec{
 		baseExecutor: base,