@lysu 2019-04-15
it's a little improve, benchmark sysbench point-get on my laptop with unistore

before this pr:

```
 sysbench --config-file=tidb-config oltp_point_select --tables=32 --threads=32 --table-size=100000 run
Unknown argument type: 0Unknown argument type: 0sysbench 1.1.0-c0bbed6 (using bundled LuaJIT 2.1.0-beta3)

Running the test with following options:
Number of threads: 32
Report intermediate results every 5 second(s)
Initializing random number generator from current time


Initializing worker threads...

Threads started!

[ 5s ] thds: 32 tps: 37581.43 qps: 37581.43 (r/w/o: 37581.43/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 10s ] thds: 32 tps: 36869.34 qps: 36869.34 (r/w/o: 36869.34/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 15s ] thds: 32 tps: 36796.11 qps: 36796.11 (r/w/o: 36796.11/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 20s ] thds: 32 tps: 36986.07 qps: 36986.07 (r/w/o: 36986.07/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 25s ] thds: 32 tps: 36634.30 qps: 36634.30 (r/w/o: 36634.30/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 30s ] thds: 32 tps: 37394.44 qps: 37394.44 (r/w/o: 37394.44/0.00/0.00) lat (ms,95%): 1.61 err/s: 0.00 reconn/s: 0.00
[ 35s ] thds: 32 tps: 37109.89 qps: 37109.89 (r/w/o: 37109.89/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 40s ] thds: 32 tps: 36963.08 qps: 36963.08 (r/w/o: 36963.08/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 45s ] thds: 32 tps: 37216.87 qps: 37216.87 (r/w/o: 37216.87/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 50s ] thds: 32 tps: 36411.83 qps: 36411.83 (r/w/o: 36411.83/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 55s ] thds: 32 tps: 34426.31 qps: 34426.31 (r/w/o: 34426.31/0.00/0.00) lat (ms,95%): 1.76 err/s: 0.00 reconn/s: 0.00
[ 60s ] thds: 32 tps: 35594.22 qps: 35594.22 (r/w/o: 35594.22/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 65s ] thds: 32 tps: 35623.28 qps: 35623.28 (r/w/o: 35623.28/0.00/0.00) lat (ms,95%): 1.70 err/s: 0.00 reconn/s: 0.00
[ 70s ] thds: 32 tps: 35443.64 qps: 35443.64 (r/w/o: 35443.64/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 75s ] thds: 32 tps: 35498.84 qps: 35498.84 (r/w/o: 35498.84/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 80s ] thds: 32 tps: 35383.51 qps: 35383.51 (r/w/o: 35383.51/0.00/0.00) lat (ms,95%): 1.70 err/s: 0.00 reconn/s: 0.00
[ 85s ] thds: 32 tps: 35476.55 qps: 35476.55 (r/w/o: 35476.55/0.00/0.00) lat (ms,95%): 1.70 err/s: 0.00 reconn/s: 0.00
[ 90s ] thds: 32 tps: 35441.08 qps: 35441.08 (r/w/o: 35441.08/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 95s ] thds: 32 tps: 35745.21 qps: 35745.21 (r/w/o: 35745.21/0.00/0.00) lat (ms,95%): 1.70 err/s: 0.00 reconn/s: 0.00
[ 100s ] thds: 32 tps: 35487.37 qps: 35487.57 (r/w/o: 35487.57/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 105s ] thds: 32 tps: 35476.37 qps: 35476.17 (r/w/o: 35476.17/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 110s ] thds: 32 tps: 35768.95 qps: 35768.95 (r/w/o: 35768.95/0.00/0.00) lat (ms,95%): 1.70 err/s: 0.00 reconn/s: 0.00
[ 115s ] thds: 32 tps: 35182.00 qps: 35182.00 (r/w/o: 35182.00/0.00/0.00) lat (ms,95%): 1.76 err/s: 0.00 reconn/s: 0.00
[ 120s ] thds: 32 tps: 34049.40 qps: 34049.40 (r/w/o: 34049.40/0.00/0.00) lat (ms,95%): 1.86 err/s: 0.00 reconn/s: 0.00
[ 125s ] thds: 32 tps: 35475.86 qps: 35475.86 (r/w/o: 35475.86/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 130s ] thds: 32 tps: 35454.40 qps: 35454.40 (r/w/o: 35454.40/0.00/0.00) lat (ms,95%): 1.70 err/s: 0.00 reconn/s: 0.00
[ 135s ] thds: 32 tps: 34809.76 qps: 34809.76 (r/w/o: 34809.76/0.00/0.00) lat (ms,95%): 1.76 err/s: 0.00 reconn/s: 0.00
[ 140s ] thds: 32 tps: 35446.99 qps: 35446.99 (r/w/o: 35446.99/0.00/0.00) lat (ms,95%): 1.76 err/s: 0.00 reconn/s: 0.00
[ 145s ] thds: 32 tps: 35519.50 qps: 35519.50 (r/w/o: 35519.50/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 150s ] thds: 32 tps: 35698.38 qps: 35698.38 (r/w/o: 35698.38/0.00/0.00) lat (ms,95%): 1.70 err/s: 0.00 reconn/s: 0.00
[ 155s ] thds: 32 tps: 35453.13 qps: 35453.13 (r/w/o: 35453.13/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 160s ] thds: 32 tps: 34945.31 qps: 34945.31 (r/w/o: 34945.31/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 165s ] thds: 32 tps: 35361.51 qps: 35361.51 (r/w/o: 35361.51/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 170s ] thds: 32 tps: 35588.72 qps: 35588.72 (r/w/o: 35588.72/0.00/0.00) lat (ms,95%): 1.70 err/s: 0.00 reconn/s: 0.00
[ 175s ] thds: 32 tps: 35391.31 qps: 35391.31 (r/w/o: 35391.31/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 180s ] thds: 32 tps: 35420.59 qps: 35420.59 (r/w/o: 35420.59/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 185s ] thds: 32 tps: 35703.86 qps: 35703.86 (r/w/o: 35703.86/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 190s ] thds: 32 tps: 35372.86 qps: 35372.86 (r/w/o: 35372.86/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 195s ] thds: 32 tps: 34779.91 qps: 34779.91 (r/w/o: 34779.91/0.00/0.00) lat (ms,95%): 1.79 err/s: 0.00 reconn/s: 0.00
[ 200s ] thds: 32 tps: 35042.69 qps: 35042.69 (r/w/o: 35042.69/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 205s ] thds: 32 tps: 35395.09 qps: 35395.09 (r/w/o: 35395.09/0.00/0.00) lat (ms,95%): 1.76 err/s: 0.00 reconn/s: 0.00
[ 210s ] thds: 32 tps: 35279.83 qps: 35279.83 (r/w/o: 35279.83/0.00/0.00) lat (ms,95%): 1.70 err/s: 0.00 reconn/s: 0.00
[ 215s ] thds: 32 tps: 34945.51 qps: 34945.51 (r/w/o: 34945.51/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 220s ] thds: 32 tps: 35566.26 qps: 35566.26 (r/w/o: 35566.26/0.00/0.00) lat (ms,95%): 1.70 err/s: 0.00 reconn/s: 0.00
[ 225s ] thds: 32 tps: 34706.71 qps: 34706.71 (r/w/o: 34706.71/0.00/0.00) lat (ms,95%): 1.76 err/s: 0.00 reconn/s: 0.00
[ 230s ] thds: 32 tps: 35301.78 qps: 35301.78 (r/w/o: 35301.78/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 235s ] thds: 32 tps: 34680.28 qps: 34680.28 (r/w/o: 34680.28/0.00/0.00) lat (ms,95%): 1.76 err/s: 0.00 reconn/s: 0.00
[ 240s ] thds: 32 tps: 34849.59 qps: 34849.59 (r/w/o: 34849.59/0.00/0.00) lat (ms,95%): 1.76 err/s: 0.00 reconn/s: 0.00
[ 245s ] thds: 32 tps: 35133.26 qps: 35133.26 (r/w/o: 35133.26/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 250s ] thds: 32 tps: 35366.43 qps: 35366.43 (r/w/o: 35366.43/0.00/0.00) lat (ms,95%): 1.70 err/s: 0.00 reconn/s: 0.00
[ 255s ] thds: 32 tps: 34666.95 qps: 34666.95 (r/w/o: 34666.95/0.00/0.00) lat (ms,95%): 1.76 err/s: 0.00 reconn/s: 0.00
[ 260s ] thds: 32 tps: 35680.87 qps: 35680.87 (r/w/o: 35680.87/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 265s ] thds: 32 tps: 34782.30 qps: 34782.30 (r/w/o: 34782.30/0.00/0.00) lat (ms,95%): 1.76 err/s: 0.00 reconn/s: 0.00
[ 270s ] thds: 32 tps: 35313.33 qps: 35313.33 (r/w/o: 35313.33/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 275s ] thds: 32 tps: 34943.42 qps: 34943.42 (r/w/o: 34943.42/0.00/0.00) lat (ms,95%): 1.76 err/s: 0.00 reconn/s: 0.00
[ 280s ] thds: 32 tps: 35509.17 qps: 35509.17 (r/w/o: 35509.17/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 285s ] thds: 32 tps: 34786.36 qps: 34786.36 (r/w/o: 34786.36/0.00/0.00) lat (ms,95%): 1.79 err/s: 0.00 reconn/s: 0.00
[ 290s ] thds: 32 tps: 35333.26 qps: 35333.26 (r/w/o: 35333.26/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 295s ] thds: 32 tps: 34202.11 qps: 34202.11 (r/w/o: 34202.11/0.00/0.00) lat (ms,95%): 1.79 err/s: 0.00 reconn/s: 0.00
[ 300s ] thds: 32 tps: 35404.56 qps: 35404.56 (r/w/o: 35404.56/0.00/0.00) lat (ms,95%): 1.70 err/s: 0.00 reconn/s: 0.00
SQL statistics:
    queries performed:
        read:                            10659452
        write:                           0
        other:                           0
        total:                           10659452
    transactions:                        10659452 (35530.85 per sec.)
    queries:                             10659452 (35530.85 per sec.)
    ignored errors:                      0      (0.00 per sec.)
    reconnects:                          0      (0.00 per sec.)

Throughput:
    events/s (eps):                      35530.8532
    time elapsed:                        300.0055s
    total number of events:              10659452

Latency (ms):
         min:                                    0.15
         avg:                                    0.90
         max:                                   34.48
         95th percentile:                        1.73
         sum:                              9592246.81

Threads fairness:
    events (avg/stddev):           333107.8750/266.53
    execution time (avg/stddev):   299.7577/0.01
```

after this pr:

```
sysbench --config-file=tidb-config oltp_point_select --tables=32 --threads=32 --table-size=100000 run   
Unknown argument type: 0Unknown argument type: 0sysbench 1.1.0-c0bbed6 (using bundled LuaJIT 2.1.0-beta3)

Running the test with following options:
Number of threads: 32
Report intermediate results every 5 second(s)
Initializing random number generator from current time


Initializing worker threads...

Threads started!

[ 5s ] thds: 32 tps: 39761.85 qps: 39762.05 (r/w/o: 39762.05/0.00/0.00) lat (ms,95%): 1.55 err/s: 0.00 reconn/s: 0.00
[ 10s ] thds: 32 tps: 39054.67 qps: 39054.47 (r/w/o: 39054.47/0.00/0.00) lat (ms,95%): 1.58 err/s: 0.00 reconn/s: 0.00
[ 15s ] thds: 32 tps: 39390.78 qps: 39390.78 (r/w/o: 39390.78/0.00/0.00) lat (ms,95%): 1.52 err/s: 0.00 reconn/s: 0.00
[ 20s ] thds: 32 tps: 39176.82 qps: 39176.82 (r/w/o: 39176.82/0.00/0.00) lat (ms,95%): 1.55 err/s: 0.00 reconn/s: 0.00
[ 25s ] thds: 32 tps: 38960.40 qps: 38960.40 (r/w/o: 38960.40/0.00/0.00) lat (ms,95%): 1.58 err/s: 0.00 reconn/s: 0.00
[ 30s ] thds: 32 tps: 38982.42 qps: 38982.42 (r/w/o: 38982.42/0.00/0.00) lat (ms,95%): 1.55 err/s: 0.00 reconn/s: 0.00
[ 35s ] thds: 32 tps: 38943.02 qps: 38943.02 (r/w/o: 38943.02/0.00/0.00) lat (ms,95%): 1.58 err/s: 0.00 reconn/s: 0.00
[ 40s ] thds: 32 tps: 39013.94 qps: 39013.94 (r/w/o: 39013.94/0.00/0.00) lat (ms,95%): 1.52 err/s: 0.00 reconn/s: 0.00
[ 45s ] thds: 32 tps: 39007.30 qps: 39007.30 (r/w/o: 39007.30/0.00/0.00) lat (ms,95%): 1.55 err/s: 0.00 reconn/s: 0.00
[ 50s ] thds: 32 tps: 39001.08 qps: 39001.08 (r/w/o: 39001.08/0.00/0.00) lat (ms,95%): 1.55 err/s: 0.00 reconn/s: 0.00
[ 55s ] thds: 32 tps: 38743.60 qps: 38743.60 (r/w/o: 38743.60/0.00/0.00) lat (ms,95%): 1.58 err/s: 0.00 reconn/s: 0.00
[ 60s ] thds: 32 tps: 38011.64 qps: 38011.84 (r/w/o: 38011.84/0.00/0.00) lat (ms,95%): 1.58 err/s: 0.00 reconn/s: 0.00
[ 65s ] thds: 32 tps: 38530.85 qps: 38530.65 (r/w/o: 38530.65/0.00/0.00) lat (ms,95%): 1.58 err/s: 0.00 reconn/s: 0.00
[ 70s ] thds: 32 tps: 38292.33 qps: 38292.33 (r/w/o: 38292.33/0.00/0.00) lat (ms,95%): 1.58 err/s: 0.00 reconn/s: 0.00
[ 75s ] thds: 32 tps: 37882.19 qps: 37882.19 (r/w/o: 37882.19/0.00/0.00) lat (ms,95%): 1.61 err/s: 0.00 reconn/s: 0.00
[ 80s ] thds: 32 tps: 38282.45 qps: 38282.45 (r/w/o: 38282.45/0.00/0.00) lat (ms,95%): 1.61 err/s: 0.00 reconn/s: 0.00
[ 85s ] thds: 32 tps: 37668.89 qps: 37668.89 (r/w/o: 37668.89/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 90s ] thds: 32 tps: 36514.50 qps: 36514.50 (r/w/o: 36514.50/0.00/0.00) lat (ms,95%): 1.70 err/s: 0.00 reconn/s: 0.00
[ 95s ] thds: 32 tps: 37990.21 qps: 37990.21 (r/w/o: 37990.21/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 100s ] thds: 32 tps: 37764.07 qps: 37764.07 (r/w/o: 37764.07/0.00/0.00) lat (ms,95%): 1.61 err/s: 0.00 reconn/s: 0.00
[ 105s ] thds: 32 tps: 37808.29 qps: 37808.29 (r/w/o: 37808.29/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 110s ] thds: 32 tps: 37877.72 qps: 37877.72 (r/w/o: 37877.72/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 115s ] thds: 32 tps: 37511.14 qps: 37511.14 (r/w/o: 37511.14/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 120s ] thds: 32 tps: 35625.77 qps: 35625.77 (r/w/o: 35625.77/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 125s ] thds: 32 tps: 37532.85 qps: 37532.85 (r/w/o: 37532.85/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 130s ] thds: 32 tps: 37421.71 qps: 37421.71 (r/w/o: 37421.71/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 135s ] thds: 32 tps: 37310.62 qps: 37310.62 (r/w/o: 37310.62/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 140s ] thds: 32 tps: 36527.00 qps: 36527.00 (r/w/o: 36527.00/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 145s ] thds: 32 tps: 37369.50 qps: 37369.50 (r/w/o: 37369.50/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 150s ] thds: 32 tps: 37364.02 qps: 37364.02 (r/w/o: 37364.02/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 155s ] thds: 32 tps: 37462.43 qps: 37462.43 (r/w/o: 37462.43/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 160s ] thds: 32 tps: 37156.94 qps: 37156.94 (r/w/o: 37156.94/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 165s ] thds: 32 tps: 36051.82 qps: 36051.82 (r/w/o: 36051.82/0.00/0.00) lat (ms,95%): 1.70 err/s: 0.00 reconn/s: 0.00
[ 170s ] thds: 32 tps: 37349.65 qps: 37349.65 (r/w/o: 37349.65/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 175s ] thds: 32 tps: 37512.56 qps: 37512.56 (r/w/o: 37512.56/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 180s ] thds: 32 tps: 37231.00 qps: 37231.00 (r/w/o: 37231.00/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 185s ] thds: 32 tps: 36047.95 qps: 36047.95 (r/w/o: 36047.95/0.00/0.00) lat (ms,95%): 1.79 err/s: 0.00 reconn/s: 0.00
[ 190s ] thds: 32 tps: 36110.44 qps: 36110.44 (r/w/o: 36110.44/0.00/0.00) lat (ms,95%): 1.79 err/s: 0.00 reconn/s: 0.00
[ 195s ] thds: 32 tps: 36256.10 qps: 36256.10 (r/w/o: 36256.10/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 200s ] thds: 32 tps: 38413.21 qps: 38413.21 (r/w/o: 38413.21/0.00/0.00) lat (ms,95%): 1.58 err/s: 0.00 reconn/s: 0.00
[ 205s ] thds: 32 tps: 36900.22 qps: 36900.22 (r/w/o: 36900.22/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 210s ] thds: 32 tps: 36651.33 qps: 36651.33 (r/w/o: 36651.33/0.00/0.00) lat (ms,95%): 1.70 err/s: 0.00 reconn/s: 0.00
[ 215s ] thds: 32 tps: 37010.85 qps: 37010.85 (r/w/o: 37010.85/0.00/0.00) lat (ms,95%): 1.61 err/s: 0.00 reconn/s: 0.00
[ 220s ] thds: 32 tps: 38518.88 qps: 38518.88 (r/w/o: 38518.88/0.00/0.00) lat (ms,95%): 1.58 err/s: 0.00 reconn/s: 0.00
[ 225s ] thds: 32 tps: 36913.58 qps: 36913.58 (r/w/o: 36913.58/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 230s ] thds: 32 tps: 36323.76 qps: 36323.76 (r/w/o: 36323.76/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 235s ] thds: 32 tps: 36686.74 qps: 36686.74 (r/w/o: 36686.74/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 240s ] thds: 32 tps: 37892.57 qps: 37892.57 (r/w/o: 37892.57/0.00/0.00) lat (ms,95%): 1.61 err/s: 0.00 reconn/s: 0.00
[ 245s ] thds: 32 tps: 37996.32 qps: 37996.32 (r/w/o: 37996.32/0.00/0.00) lat (ms,95%): 1.61 err/s: 0.00 reconn/s: 0.00
[ 250s ] thds: 32 tps: 36980.85 qps: 36980.85 (r/w/o: 36980.85/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 255s ] thds: 32 tps: 37615.03 qps: 37615.03 (r/w/o: 37615.03/0.00/0.00) lat (ms,95%): 1.61 err/s: 0.00 reconn/s: 0.00
[ 260s ] thds: 32 tps: 37844.23 qps: 37844.43 (r/w/o: 37844.43/0.00/0.00) lat (ms,95%): 1.61 err/s: 0.00 reconn/s: 0.00
[ 265s ] thds: 32 tps: 36999.00 qps: 36998.80 (r/w/o: 36998.80/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 270s ] thds: 32 tps: 37705.31 qps: 37705.31 (r/w/o: 37705.31/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 275s ] thds: 32 tps: 37142.45 qps: 37142.45 (r/w/o: 37142.45/0.00/0.00) lat (ms,95%): 1.61 err/s: 0.00 reconn/s: 0.00
[ 280s ] thds: 32 tps: 37702.53 qps: 37702.53 (r/w/o: 37702.53/0.00/0.00) lat (ms,95%): 1.61 err/s: 0.00 reconn/s: 0.00
[ 285s ] thds: 32 tps: 37129.52 qps: 37129.52 (r/w/o: 37129.52/0.00/0.00) lat (ms,95%): 1.61 err/s: 0.00 reconn/s: 0.00
[ 290s ] thds: 32 tps: 37163.84 qps: 37163.84 (r/w/o: 37163.84/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 295s ] thds: 32 tps: 37406.37 qps: 37406.37 (r/w/o: 37406.37/0.00/0.00) lat (ms,95%): 1.61 err/s: 0.00 reconn/s: 0.00
[ 300s ] thds: 32 tps: 37786.79 qps: 37786.79 (r/w/o: 37786.79/0.00/0.00) lat (ms,95%): 1.58 err/s: 0.00 reconn/s: 0.00
SQL statistics:
    queries performed:
        read:                            11296519
        write:                           0
        other:                           0
        total:                           11296519
    transactions:                        11296519 (37654.39 per sec.)
    queries:                             11296519 (37654.39 per sec.)
    ignored errors:                      0      (0.00 per sec.)
    reconnects:                          0      (0.00 per sec.)

Throughput:
    events/s (eps):                      37654.3918
    time elapsed:                        300.0053s
    total number of events:              11296519

Latency (ms):
         min:                                    0.13
         avg:                                    0.85
         max:                                   32.76
         95th percentile:                        1.64
         sum:                              9592685.32

Threads fairness:
    events (avg/stddev):           353016.2188/405.50
    execution time (avg/stddev):   299.7714/0.00

```

mainly cut off cost:

![Screenshot from 2019-04-15 12-28-53](https://user-images.githubusercontent.com/528332/56107709-5b10fe80-5f7a-11e9-93c0-8f9ee2e76122.png)

sysbench's point-get will got a very simple executor-tree, it will be more usefule iff executor-tree be more complicate